- Added an additional constructor to the VirtualStreamReader: Now it is possible to create the VirtualStreamReader with a MemoryStream. - Changed the console outputs to traceoutputs - Added a console program with helpoutput - Changed some Biffrecords - Changed some Exceptionoutputs - Changed the SST extraction - Bugfixes etc.

+++ src/Spreadsheet/XlsFileFormat/BiffRecords/MULBLANK.cs
@@ -73,7 +73,7 @@ namespace DIaLOGIKa.b2xtranslator.Spreadsheet.XlsFileFormat.BiffRecords
             this.ixfe = new List<UInt16>();
 
             // count records - 6 standard non variable values !!! 
-            int count = (this.Length - 6) / 2;
+            int count = (int)(this.Length - 6) / 2;
             this.rw = reader.ReadUInt16();
             this.colFirst = reader.ReadUInt16();
             for (int i = 0; i < count; i++)
